pyproject.toml
[project]
name = "fizzbuzz-forever"
version = "0.1.0"
description = "Train a small LLM to be a fizzbuzz agent via SFT with tool calling"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "trl>=0.29.0",
    "transformers",
    "datasets",
    "tensorboard",
    "setuptools<81",
    "pyyaml",
    "accelerate",
    "num2words>=0.5.14",
    "jmespath",
]

[project.scripts]
fizzbuzz-train = "fizzbuzz_forever.train:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fizzbuzz_forever"]
